fix validation error crashing ghost
fixes fatal error that results in status 500 errors when theme includes a file that ends in [^.].hbs
also fixes that theme fails to activate of theme includes files ending in [^.].hbs

Example: 
Create new theme, and create new file amp.__hbs. 
Observe that ghost crashes if theme is already in use, else theme can't be activated (not error is displayed).
Upon inspecting ghost logs, observe fatal crash in line 26 (`if (themeFile.content.match(check.regex))`) "Can't read match of undefined".
This is because content only exists for actual hbs templates and a file ".__hbs" has not been processed as a template.

Fixing the regex fixes this issue by properly detecting if a file is actuall an ".hbs"-file, not a "<any character>hbs"-file (the `.` character needs to be scaped in js regex, else it matches any character).