Fixed detection of custom theme settings in get helper filter attributes (#636)
ref https://linear.app/ghost/issue/ONC-1249/theme-scan-shows-error-on-custom-theme-setting-but-the-setting-works
closes #464

Problem:
There is a bug in gscan's "GS100-NO-UNUSED-CUSTOM-THEME-SETTING" check, which causes a false positive failure if a custom theme setting is only used within the filter attribute of a get helper:

{{#get "authors" limit="9" filter="slug:[{{@custom.author_slugs}}]"}}

Cause:
Handlebars itself doesn't parse the StringLiteral arguments, so the @custom.author_slugs custom theme setting from the example above is not added to the AST directly, and therefore it's not detected as "used" in the theme.

Ghost itself does some additional parsing here, specifically for the {{#get}} helper, and specifically for the filter="" attribute only, so the example above works fine in Ghost, but results in a gscan error.

Solution:
A naive solution would be to use a regex match on all files looking for @custom.*, but this would largely defeat the purpose of using the AST at all, and we'd run the risk of false negatives if @custom.author_slugs were included in a comment.

Instead, this narrows the scope to scan through all {{#get}} helpers in the theme, then do a regex check on only the filter attributes StringLiteral value.